change order of pin operations in QTRSensorsRC::readPrivate to match comments
This resolves the inconsistency pointed out in
https://github.com/pololu/qtr-sensors-arduino/pull/6.  As I described in
my comment on that PR, we probably intentionally put the digitalWrite
call first to make the pin pull high before driving high (so it was
arguably the comments that were reversed, not the function calls).
Swapping the function calls like I'm now doing changes the behavior so
the pin drives LOW briefly before driving high, but we looked at this
carefully and decided it would have no impact on the sensor results. The
argument for changing it this way is that it allows the library to work
properly on platforms where calling pinMode(x, OUTPUT) does not make the
pin output a value previously set with digitalWrite.